Moved notification into its own component class for readability
